refactor: '@/shared/hooks'에서 useFetchHooks를 통합하여 코드 정리
