+++ src/app/App.jsx
@@ -3,8 +3,7 @@ import { BrowserRouter } from 'react-router-dom';
 
 import AppRouter from '@/app/router';
 import { store } from '@/redux';
-import ScrollToTop from '@/shared/hooks/scrollToTop';
-import AuthAndUserInfo from '@/shared/hooks/useAuthAndUserInfo';
+import { AuthAndUserInfo, ScrollToTop } from '@/shared/hooks';
 
 const App = () => {
   return (
