+++ src/shared/hooks/ScrollToTop.ts
@@ -11,4 +11,4 @@ const ScrollToTop = () => {
   return null;
 };
 
-export default ScrollToTop;
+export { ScrollToTop };
