+++ src/shared/hooks/index.ts
@@ -1,0 +1,4 @@
+export * from './scrollToTop';
+export * from './useAuthAndUserInfo';
+export * from './useFetchHooks';
+export * from './useIntersectionObserver';
